Moved scene serialization to editor
- Renamed GrappleCommon to GrappleCore

+++ Grapple/src/Grapple/Platform/OpenGL/OpenGLFrameBuffer.cpp
@@ -1,6 +1,6 @@
 #include "OpenGLFrameBuffer.h"
 
-#include "Grapple/Core/Assert.h"
+#include "GrappleCore/Assert.h"
 
 #include <glad/glad.h>
 #include <glm/gtc/type_ptr.hpp>
@@ -144,9 +144,9 @@ namespace Grapple
         for (size_t i = 0; i < m_ColorAttachments.size(); i++)
         {
             if (IsDepthFormat(m_Specifications.Attachments[i].Format))
-                AttachDethTexture(i);
+                AttachDethTexture((uint32_t)i);
             else
-                AttachColorTexture(i);
+                AttachColorTexture((uint32_t)i);
         }
 
         Grapple_CORE_ASSERT(m_ColorAttachments.size() <= 3);
